jlb - Notifications, Config, Animations
Doing mashup with notifications team. Added Stomp client. Updated the chat view to request the alert channel through the notification REST api and then subscribe to it.

Created a config class. Right now it stores the backend uri and the notification uri. It's also used to disable notifications and/or use the mock poker engine

Had to add a method to the player model to reformat the backend's card data from an object to a string.

Added more animations:
- On showdown, the cards of the winners are flipped and the chips image is moved to that player(s)
- On start of new hand, the player cards fade out

Testing incoming data queue. We want to delay incoming data between rounds (or other states). We do this by moving any data coming from web sockets into a queue and then for each round we would keep track of all the animations and only move onto the next round (shifting off the data queues) when all animations are done. Looks good on paper.