jlb - Scenarios and Effects
Created a Scenarios namespace to organize the scenarios that we need to test against. A proposed workflow would be to create a scenario to test a bug. This way we can create a library of scenarios to test the UI against.

Updated the transition from river to deal. There was a bug found that didn't fade out the buttons. Also added an effect to show a positive or negative change in a player's chips.